Updated README to reflect MacOS support

+++ include/cognate.c
@@ -35,10 +35,8 @@ static void init(int argc, char** argv)
   stack_start = &a;
   getrlimit(RLIMIT_STACK, &stack_max);
   // Get executable path stuff.
-  // TODO readlink probably doesn't work on macOS.
-  // TODO read first argument for filepath relative to cwd instead.
-#ifdef __APPLE__
-  int bufsize = PATH_MAX;
+#ifdef __APPLE__ // '/proc/self/exe' doesn't exist on macos
+  short bufsize = PATH_MAX;
   _NSGetExecutablePath(file_name_buf, &bufsize);
 #else
   readlink("/proc/self/exe", file_name_buf, PATH_MAX);
